requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
SQLAlchemy==2.0.25
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0